beberapa update

+++ database/migrations/2022_04_24_145652_create_menu_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('Menu', function (Blueprint $table) {
             $table->id();
             $table->string('menuname');
-            $table->integer('price');
+            $table->decimal('price',7,0);
             $table->string('image');
             $table->timestamps();
         });
